Avoid duplicating blocks when inserting signal checks
Now that functions have multiple entry points, we must be more careful
about how we handle things that want to modify what happens at the
"beginning" of a function.  In this instance, we could run into
situations where we would add signal checks to a block more than once
since there was no check to make sure that loopForestSteensgaard gave us
unique blocks that we had not already visited before.  If a node is
visited twice by insertSignalChecks, it will be duplicated (since
insertSignalChecks creates a new block with the label of the header
block and adds this new block to the function).  A simple check to see
if we've already marked a label as a header indicates if we've already
added a signal check to it.