case-entries: entryInfo doesn't need an option type
When I initially wrote this pass, I though that it may be useful for
entryInfo to take in a 'Con.t option' type: if NONE was passed in then
the original 'FunctionEntry.t' would be returned.

The one callee of this function doesn't seem to need that behavior (the
callee knows the default entry already).  There doesn't seem to be
anything lost here by not using an option type.