case-entries: attempt to add entries to multi-arg entries
Currently, we can only create a new entry point if the existing argument is a
single variable that is a constructed datatype.  It made the design simpler,
in the new entry just swap out the constructor's arguments with your own, but
left some opportunities for adding new entry points unexplored.

This change seeks to rectify that, by removing the contructed datatype from
the argument and adding in the arguments that were used to construct the
datatype.  As in a previous change, we re-create the constructed datatype in
the new entry in case it's referred to directly (otherwise dead code
elimination removes it for us).

oldcaller():
    ...
    convar = Constructor(carg1, carg2, carg3)
    ...
    callee@oldentry(convar, "apples")

callee:
oldentry(convar, fruit):
    ...
    <stuff>
    ...

transforms to:

oldcaller():
    ...
    callee@newentry(carg1, carg2, carg3, "apples")

callee:
oldentry(convar, fruit):
    ...
    <stuff>
    ...
newentry(carg1, carg2, carg3, fruit)
    ...
    convar = Constructor(carg1, carg2, carg3)
    ...

Currently there seems to be some issue in the SSA restore transformation
related to renaming the arguments in the new entries.  It appears that if
multiple entry points start by using the same argument names, the variable
renamer renames all of the arg names; but when updating all references to the
original variables, it only uses names from the newest entry point it touched,
instead of the one we know will be in scope.

Say we end up with this before passing it to restore.

callee:
entry1(carg1, carg2, fruit)
    ...
    newvar = stuff(carg1, carg2)
    ...
entry2(carg1, carg2, fruit)
    ...
    newvar = stuff(carg1, carg2)
    ...

After variable renaming in the restore transformation, we can end up with a
function that looks like this:

callee:
entry1(x_10, x_11, x_12)
    ...
    newvar = stuff(x_13, x_14)
    ...
entry2(x_13, x_14, x_15)
    ...
    newvar = stuff(x_13, x_14)
    ...

x_13 isn't in scope for entry1, so this transformation doesn't make it past
the typechecker.

We notices this when running the tyan and vliw benchmarks.